Updated handling of reference frame properties in the enzo-p frontend to reflect updates in the frame tracking scheme in enzo-e.
Note: The origin_offset will be wrong if processing outputs written by versions of enzo-e that employed the older version of the frame tracking scheme. However, that version of the scheme was buggy (it only correctly saved origin_offset if the update schedule was based off of schedules - if the updates were based off of simulation time, then the origin_offsets were not appropriately saved). Moreover, the older version of the scheme was never added to the master branch.